Add user authentication and CRUD

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel")
 
 // @desc Get all contact
 // @route GET /api/contacts
-// @acces public
+// @acces private
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 // @desc Create new contact
 // @route POST /api/contacts
-// @acces public
+// @acces private
 
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request bosy is: ", req.body);
@@ -25,14 +25,15 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     })
     res.status(201).json(contact);
 });
 
 // @desc Create new contact
 // @route GET /api/contacts/:id
-// @acces public
+// @acces private
 
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -45,13 +46,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc Update new contact
 // @route PUT /api/contacts/:id
-// @acces public
+// @acces private
 
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contact");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -65,13 +71,18 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete new contact
 // @route DELETE /api/contacts/:id
-// @acces public
+// @acces private
 
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user contact");
     }
 
     await Contact.findByIdAndDelete(req.params.id);
